connected api with the database

+++ pruebaMysql.sql
@@ -1,10 +1,11 @@
 CREATE DATABASE  IF NOT EXISTS assestment;
 USE assestment;
+
 /*Se crean las tablas de countries y languages*/
-CREATE TABLE IF NOT EXISTS countries (id int AUTO_INCREMENT NOT NULL PRIMARY KEY, name VARCHAR (50),
+CREATE TABLE IF NOT EXISTS countries (id LONG AUTO_INCREMENT NOT NULL PRIMARY KEY, name VARCHAR (50),
        capital VARCHAR(50), population int, languageId int);
 
-CREATE TABLE IF NOT EXISTS languages (id int AUTO_INCREMENT NOT NULL PRIMARY KEY, name VARCHAR (50));	
+CREATE TABLE IF NOT EXISTS languages (id long AUTO_INCREMENT NOT NULL PRIMARY KEY, name VARCHAR (50));	
 
 /*
 INSERT INTO countries (name,capital,population,languageId) VALUES ("Mexico", "Ciudad de Mexico",1289000,0);
